style(bb-test): rename wait-assert recipe module to its primary export
